Testing: Remove test_codeeditor_1.py
Move its tests to other files instead
